camera 3_4: Fix compilation errors
Compile Error:
hardware/libhardware/modules/camera/3_4/camera.cpp:502:9: error: non-constant-expression cannot be narrowed from type 'std::__1::vector<camera3_stream_buffer, std::__1::allocator<camera3_stream_buffer> >::size_type' (aka 'unsigned long') to 'uint32_t' (aka 'unsigned int') in initializer list [-Wc++11-narrowing]
        request->output_buffers.size(),
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Compile Error:
hardware/libhardware/modules/camera/3_4/v4l2_gralloc.cpp:168:15: error: use of undeclared identifier 'EINVAL'
      return -EINVAL;

Test: Add to device.mk
        USE_CAMERA_V4L2_HAL := true
        PRODUCT_PACKAGES += camera.v4l2
      mma

Merged-In: Ia6bbd6bac06540214d2c6167a8826f3f4c946e57

+++ modules/camera/3_4/request_tracker_test.cpp
@@ -35,7 +35,7 @@ class RequestTrackerTest : public Test {
     stream2_.max_buffers = 3;
     dut_.reset(new RequestTracker());
     streams_ = {&stream1_, &stream2_};
-    camera3_stream_configuration_t config{streams_.size(), streams_.data(), 0};
+    camera3_stream_configuration_t config{static_cast<uint32_t>(streams_.size()), streams_.data(), 0};
     dut_->SetStreamConfiguration(config);
   }
